+++ src/powernse/reading/__init__.py
@@ -2,6 +2,7 @@
 
 from powernse.reading.bhavcopy import BhavcopyReader
 from powernse.reading.corporate import CorporateActionReader
+from powernse.reading.deals import DealsReader, SecbanReader
 from powernse.reading.delivery import DeliveryReader
 from powernse.reading.futures import FoReader
 from powernse.reading.index import IndexReader
@@ -10,8 +11,10 @@ from powernse.reading.snapshots import SnapshotReader
 __all__ = [
     "BhavcopyReader",
     "CorporateActionReader",
+    "DealsReader",
     "DeliveryReader",
     "FoReader",
     "IndexReader",
+    "SecbanReader",
     "SnapshotReader",
 ]